Refactor part about Representation in Board, Piece, BoardTest, PieceTest class Representation enum의 변경에 따라 Board, Piece, BoardTest, PieceTest class 리팩토링
