+++ src/main/java/net/eno/pieces/Piece.java
@@ -5,14 +5,9 @@ public class Piece {
     private final String color;
     private final char representation;
 
-    public Piece() {
-        this.color = Color.WHITE.toString();
-        this.representation = Representation.WHITE.getPawn();
-    }
-
     public Piece(Color color, Representation representation) {
         this.color = color.toString();
-        this.representation = representation.getPawn();
+        this.representation = representation.get(color);
     }
 
     public String getColor() {
